+++ framehud-metrics/src/androidTest/kotlin/com/timkrest/framehud/BlankActivity.kt
@@ -1,0 +1,5 @@
+package com.timkrest.framehud
+
+import android.app.Activity
+
+class BlankActivity : Activity()