+++ framehud-metrics/src/main/kotlin/com/timkrest/framehud/FrameHudPanel.kt
@@ -1,0 +1,14 @@
+package com.timkrest.framehud
+
+import androidx.annotation.MainThread
+
+@MainThread
+@InternalFrameHudApi
+public interface FrameHudPanel {
+
+    public fun onConfigChanged()
+
+    public fun onScreenFocused()
+
+    public fun onScreenLost()
+}