+++ framehud/src/main/kotlin/com/timkrest/framehud/internal/LogTag.kt
@@ -1,5 +1,0 @@
-package com.timkrest.framehud.internal
-
-import com.timkrest.framehud.LogcatEventListener
-
-internal const val LOG_TAG = LogcatEventListener.TAG